+++ api/index.js
@@ -1,61 +1,6 @@
-// index.js
-const express = require('express');
-const cattracks = require('../cattracks.js');
-const cattracksfull = require('../cattracksfull.js');
-const cattracks_route = require('../cattracks_route.js');
-const routeplanner = require('./routeplanner.js');
+// api/index.js (or your preferred file name)
+const app = require('./index.js');
 
-// instantiate the express server
-const app = express();
-
-// used to get post events as JSON objects correctly
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-	
-app.get('/', function (req, res) {
-    res.send('Hi there! You should not be seeing me. 😅');
-  })
-
-app.get('/cattracks', async function (req, res) {
-    await cattracks.getRouteData();
-    await cattracks.getSchedule();
-    await cattracks.getStops();
-    console.log(req.headers);
-    console.log(req.query);
-    res.json(cattracks.buildHome(req.query));
-    console.log("ALL DONE! 🎉")
-})
-
-app.get('/cattracksfull', async function (req, res) {
-    await cattracksfull.getRouteData();
-    await cattracksfull.getSchedule();
-    await cattracksfull.getStops();
-    console.log(req.headers);
-    console.log(req.query);
-    res.json(cattracksfull.buildHome(req.query));
-    console.log("ALL DONE! 🎉")
-})
-
-
-app.post('/cattracks/routeplanner', async function (req, res) {
-    console.log(req.headers);
-    console.log(req.query);
-    console.log(req.body);
-
-    await routeplanner.getRouteData();
-    await routeplanner.getSchedule();
-    await routeplanner.getStops();
-
-    res.json(routeplanner.buildRoutePlan(req.body));
-})
-
-app.get('/cattracks/route/:id', async function (req, res) {
-    await cattracks_route.getRouteData();
-    await cattracks_route.getSchedule();
-    console.log(req.headers);
-    let xmJson = cattracks_route.buildRouteInformation(req.params.id);
-    res.json(xmJson);
-    console.log("ALL DONE! 🎉")
-})
-
-module.exports = app;
+module.exports = async (req, res) => {
+  await app(req, res); 
+};
